consumer/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
# kafka
confluent-kafka==2.5.3
# apache
apache-beam==2.59.0
# cassandra
cassandra-driver==3.29.2
# prometheus
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0